Added ability for players to spend credits to join a 'paid online' game - this is to simulate a payment system that is out pf scope for this project currently so simulated by giving users 100 credits when registering for demo purposes. Includes: - PaymentType.PAID_ONLINE added back - it was previously removed when payment processing was descoped, but now re-added since credits act as a proxy for it and better for demoing features and having an actual checkout process for demoing concurrency features better - Credit.amount changed from Integer to BigDecimal to reflect money ie. cost of a game could be in pounds and pence - CreditRepository.sumAmountByUser query added to get a user's credit balance summing the amount column in the Credit table filtered by user - BookingService.getJoinBlockReason now blocks joining a PAID_ONLINE game if the player's credit balance is less than the game price - BookingService.bookSlot deducts the game price from the player's credits as a negative Credit record when booking a PAID_ONLINE game - UserService.SIGNUP_CREDIT_AMOUNT updated to BigDecimal to match Entity change - BookingServiceTest and UserServiceTest updated for JUnit tests

+++ src/main/java/com/dpapie01/distributed_booking_system/service/UserService.java
@@ -18,11 +18,13 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.math.BigDecimal;
+
 @RequiredArgsConstructor
 @Service
 public class UserService {
 
-    private static final int SIGNUP_CREDIT_AMOUNT = 100;
+    private static final BigDecimal SIGNUP_CREDIT_AMOUNT = BigDecimal.valueOf(100);
 
     private final UserRepository userRepository;
     private final LocationRepository locationRepository;
